# Split the span start/end benchmarks and test start with links and attributes