@dmathieu 2024-06-27
I have also extended the benchmark to include links and attributes. Current results:

```
pkg: go.opentelemetry.io/otel/sdk/trace
BenchmarkTraceStart/with_a_simple_span-10                1822948               671.1 ns/op           704 B/op         14 allocs/op
BenchmarkTraceStart/with_several_links-10                1500474               840.0 ns/op           880 B/op         15 allocs/op
BenchmarkTraceStart/with_attributes-10                   1349312               869.6 ns/op           960 B/op         16 allocs/op
PASS
```